+++ program/Cargo.toml
@@ -34,14 +34,14 @@ spl-transfer-hook-interface = "0.9.0"
 thiserror = "2.0.12"
 
 # Should depend on the next crate version after 7.0.0 when https://github.com/solana-program/token-2022/pull/233 is deployed
-spl-token-2022 = { git = "https://github.com/solana-program/token-2022", rev = "612ef3363fe17d6e883617989f58b45dbff92415", features = ["no-entrypoint"] }
+spl-token-2022 = { git = "https://github.com/solana-program/token-2022", rev = "d4e342c61dff58a5782f704679323fda39b7113d", features = ["no-entrypoint"] }
 
 [dev-dependencies]
 mollusk-svm = "0.1.1"
 mollusk-svm-programs-token = "0.1.1"
 spl-tlv-account-resolution = "0.9.0"
 solana-account = "2.2.1"
-solana-system-program = "2.2.0"
+solana-sdk-ids = "2.2.1"
 test-transfer-hook = { path = "tests/helpers/test-transfer-hook", features = ["no-entrypoint"] }
 
 [lib]
