+++ program/src/lib.rs
@@ -5,7 +5,6 @@
 mod entrypoint;
 pub mod error;
 pub mod instruction;
-pub mod offchain;
 pub mod processor;
 pub mod state;
