+++ program/tests/helpers/test-transfer-hook/Cargo.toml
@@ -6,10 +6,8 @@ publish = false
 
 [features]
 no-entrypoint = []
-test-sbf = []
 
 [dependencies]
-bytemuck = { version = "1.21.0", features = ["derive"] }
 solana-account-info = "2.1.13"
 solana-program-entrypoint = "2.1.13"
 solana-program-error = "2.1.13"
@@ -17,9 +15,6 @@ solana-pubkey = "2.1.13"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
 
 [lints]
 workspace = true
