+++ program/tests/helpers/test-transfer-hook/src/lib.rs
@@ -2,6 +2,5 @@
 
 pub mod entrypoint;
 pub mod processor;
-pub mod state;
 
 solana_pubkey::declare_id!("HookipyvZYC9h3AG1x7qfdXmE3TnqvqV9ZK43ErZfBrW");
