+++ program/tests/helpers/test-transfer-hook/src/processor.rs
@@ -1,7 +1,6 @@
 //! Program state processor
 
 use {
-    crate::state::Counter,
     solana_account_info::{next_account_info, AccountInfo},
     solana_program_error::ProgramResult,
     solana_pubkey::Pubkey,
@@ -23,11 +22,7 @@ pub fn process_instruction(
 
     // Increment the counter
     let mut counter_data = counter_account.try_borrow_mut_data()?;
-    let counter_size = std::mem::size_of::<Counter>();
-    let counter_slice = &mut counter_data[..counter_size];
-
-    let counter = bytemuck::from_bytes_mut::<Counter>(counter_slice);
-    counter.count = counter.count.checked_add(1).unwrap();
+    counter_data[0] = counter_data[0].checked_add(1).unwrap();
 
     Ok(())
 }
