--- program/tests/helpers/test-transfer-hook/src/state.rs
@@ -1,7 +1,0 @@
-use bytemuck::{Pod, Zeroable};
-
-#[repr(transparent)]
-#[derive(Clone, Copy, Debug, Default, PartialEq, Pod, Zeroable)]
-pub struct Counter {
-    pub count: u64,
-}
