+++ program/tests/test_unwrap.rs
@@ -16,7 +16,6 @@ use {
         pod::{PodAccount, PodMint},
     },
     spl_token_wrap::{error::TokenWrapError, get_wrapped_mint_address, get_wrapped_mint_authority},
-    test_transfer_hook::state::Counter,
 };
 
 pub mod helpers;
@@ -317,8 +316,6 @@ fn test_unwrap_with_transfer_hook() {
     );
 
     // Verify counter was incremented
-    let counter_data = unwrap_result.extra_accounts[0].clone().account.data;
-    let counter_slice = &counter_data[..std::mem::size_of::<Counter>()];
-    let counter = bytemuck::from_bytes::<Counter>(counter_slice);
-    assert_eq!(counter.count, 1)
+    let count = unwrap_result.extra_accounts[0].clone().account.data[0];
+    assert_eq!(count, 1)
 }
